Add file path as tooltip in Documents menu
Same functionality as in the Documents Tool or the tabs
